LayoutTests/platform/gtk/http/tests/security/video-cross-origin-accessfailure-expected.txt

CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
EVENT(error)
END OF TEST
